Move general-purpose utils from demoscene.utils to common.utils
