+++ common/utils/text.py
