+++ demoscene/context_processors.py
@@ -1,8 +1,8 @@
 from django.conf import settings
 from django.contrib.auth.forms import AuthenticationForm
 
-from demoscene.utils.accounts import is_login_banned
-from demoscene.utils.ajax import request_is_ajax
+from common.utils.accounts import is_login_banned
+from common.utils.ajax import request_is_ajax
 from search.forms import SearchForm
