+++ demoscene/models.py
@@ -17,8 +17,8 @@ from django.utils.functional import cached_property
 from django.utils.safestring import mark_safe
 from unidecode import unidecode
 
-from demoscene.utils import groklinks
-from demoscene.utils.text import generate_search_title
+from common.utils import groklinks
+from common.utils.text import generate_search_title
 from lib.lockable import Lockable
 from lib.prefetch_snooping import ModelWithPrefetchSnooping
 from lib.strip_markup import strip_markup
