+++ forums/views.py
@@ -10,7 +10,7 @@ from django.utils.decorators import method_decorator
 from django.views.generic import TemplateView
 from read_only_mode import writeable_site_required
 
-from demoscene.utils.pagination import PaginationControls
+from common.utils.pagination import PaginationControls
 from demoscene.views.generic import AjaxConfirmationView
 from forums.forms import NewTopicForm, ReplyForm
 from forums.models import Post, Topic
