+++ janeway/importing.py
@@ -4,7 +4,7 @@ from django.db import transaction
 from django.db.models import Max
 
 from demoscene.models import Releaser, ReleaserExternalLink
-from demoscene.utils.text import strip_music_extensions
+from common.utils.text import strip_music_extensions
 from platforms.models import Platform
 from productions.models import Production, ProductionLink, ProductionType
