+++ janeway/matching.py
@@ -3,7 +3,7 @@ from collections import defaultdict
 from django.db.models import Q
 
 from demoscene.models import Releaser, ReleaserExternalLink
-from demoscene.utils.text import generate_search_title, strip_music_extensions
+from common.utils.text import generate_search_title, strip_music_extensions
 from janeway.importing import import_release
 from janeway.models import AuthorMatchInfo
 from janeway.models import Release as JanewayRelease
