+++ productions/management/commands/fetch_embed_data.py
@@ -1,6 +1,6 @@
 from django.core.management.base import BaseCommand
 
-from demoscene.utils.groklinks import EMBEDDABLE_PRODUCTION_LINK_TYPES
+from common.utils.groklinks import EMBEDDABLE_PRODUCTION_LINK_TYPES
 from productions.models import ProductionLink
 from productions.tasks import fetch_production_link_embed_data
