+++ productions/management/commands/refetch_embed_data.py
@@ -2,7 +2,7 @@ import datetime
 
 from django.core.management.base import BaseCommand
 
-from demoscene.utils.groklinks import EMBEDDABLE_PRODUCTION_LINK_TYPES
+from common.utils.groklinks import EMBEDDABLE_PRODUCTION_LINK_TYPES
 from productions.models import ProductionLink
 from productions.tasks import fetch_production_link_embed_data
