+++ search/views.py
@@ -4,7 +4,7 @@ from django.shortcuts import render
 
 from bbs.models import BBS
 from demoscene.models import Releaser
-from demoscene.utils.text import generate_search_title
+from common.utils.text import generate_search_title
 from parties.models import Party
 from productions.models import Production, Screenshot
 from search.forms import SearchForm
