+++ tournaments/models.py
@@ -7,7 +7,7 @@ from django.db import models
 from django.utils.functional import cached_property
 
 from demoscene.models import ExternalLink, Nick
-from demoscene.utils import groklinks
+from common.utils import groklinks
 from parties.models import Party
 from screenshots.models import PILConvertibleImage, ThumbnailMixin
 from screenshots.processing import upload_to_s3
